Treesame-Commit-Id: 03501e5d66f89c56ad1fdecc8f91f6f9d9a47b90

+++ lib/tasks/js.rake
@@ -6,13 +6,15 @@ namespace :js do
   task :build_client_apps do
     require 'config/initializers/client_app_symlinks'
 
+    npm_install = ENV["COMPILE_ASSETS_NPM_INSTALL"] != "0"
+
     Dir.glob('./client_apps/*/').each do |app_dir|
       app_name = File.basename(app_dir)
 
       Dir.chdir(app_dir) do
         puts "Building client app '#{app_name}'"
 
-        if File.exists?('./package.json')
+        if npm_install && File.exists?('./package.json')
           output = system 'yarn install --pure-lockfile || yarn install --pure-lockfile --network-concurrency 1'
           unless $?.exitstatus == 0
             puts "INSTALL FAILURE:\n#{output}"
